Futher adjustment in project structure, imports, etc.
