+++ reinforcement_learning/mdp.py
@@ -1,7 +1,15 @@
+#BEGIN--------------------PROJECT-ROOT-PATH-APPENDING-------------------------#
+import sys, os
+REL_PROJECT_ROOT_PATH = "./../"
+ABS_FILE_DIR = os.path.dirname(os.path.abspath(__file__))
+ABS_PROJECT_ROOT_PATH = os.path.normpath(os.path.join(ABS_FILE_DIR, REL_PROJECT_ROOT_PATH))
+sys.path.append(ABS_PROJECT_ROOT_PATH)
+#-------------------------PROJECT-ROOT-PATH-APPENDING----------------------END#
+
 from graphviz import Digraph
 import numpy as np
 
-from auxiliary_utilities import linear_map
+from reinforcement_learning.auxiliary_utilities import linear_map
 
 
 class MDP:
