+++ training_platform/clients/basic_player_clients/rl_player_client.py
@@ -1,8 +1,15 @@
-from common import *
-import os, sys
-sys.path.append(os.path.join(os.path.dirname(__file__), "../../reinforcement_learning/agents/"))
-from new_basic_agent import BasicAgent
-from service import GameManager, MatchMaker
+#BEGIN--------------------PROJECT-ROOT-PATH-APPENDING-------------------------#
+import sys, os
+REL_PROJECT_ROOT_PATH = "./../../../"
+ABS_FILE_DIR = os.path.dirname(os.path.abspath(__file__))
+ABS_PROJECT_ROOT_PATH = os.path.normpath(os.path.join(ABS_FILE_DIR, REL_PROJECT_ROOT_PATH))
+sys.path.append(ABS_PROJECT_ROOT_PATH)
+#-------------------------PROJECT-ROOT-PATH-APPENDING----------------------END#
+
+from reinforcement_learning.agents.basic_agent import BasicAgent
+from training_platform.server.common import *
+from training_platform.server.service import GameManager, MatchMaker
+
 import signal
 import sys
 from logger import Logger
